% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me.R
\name{plot_me}
\alias{plot_me}
\title{Plot marginal effects from two-way interactions in linear regressions}
\source{
Inspired by:
\url{http://www.carlislerainey.com/2013/08/27/creating-marginal-effect-plots-for-linear-regression-models-in-r/}
}
\usage{
plot_me(obj, term1, term2, fitted2, ci = 90)
}
\arguments{
\item{obj}{fitted model object from \code{lm}.}

\item{term1}{character string of the first constitutive term of the
interaction's variable name.}

\item{term2}{character string of the second constitutive term of the
interaction's variable name.}

\item{fitted2}{numeric vector of fitted values of \code{term2} to plot for.
If unspecified, then all unique observed values are used.}

\item{ci}{numeric. Either 90 or 95 indicating that the 90 or 95% confidence
interval should be used.}
}
\value{
a \code{gg} class ggplot2 object
}
\description{
Plot marginal effects from two-way interactions in linear regressions
}

